refactor: applied feedback from Josh
