+++ optimize/backend/src/main/java/io/camunda/optimize/service/dashboard/InstantPreviewDashboardService.java
@@ -7,7 +7,7 @@
  */
 package io.camunda.optimize.service.dashboard;
 
-import static io.camunda.optimize.TomcatConfig.EXTERNAL_SUB_PATH;
+import static io.camunda.optimize.OptimizeTomcatConfig.EXTERNAL_SUB_PATH;
 import static io.camunda.optimize.dto.optimize.ReportConstants.ALL_VERSIONS;
 import static io.camunda.optimize.dto.optimize.query.dashboard.InstantDashboardDataDto.INSTANT_DASHBOARD_DEFAULT_TEMPLATE;
 import static io.camunda.optimize.tomcat.OptimizeResourceConstants.STATIC_RESOURCE_PATH;
